Replace prefix Safe by CloseShield #1449
